Update video_editors_render_script.py
Please find this proposal to allow the script to render a scene named passed as argument on the blender command line.
As compositing can be mandatory to some users, I had the idea to make the python script more flexible. If no argument is provided, the default "Scene" name is considered. If an argument is provided after on the command line after -- then it is given to the script and will be our target Scene.
There are a few consequences
- new lines to handle the argument
- there cannot be any control regarding the lack of VSE strips because VSE MUST BE EMPTY if we want a compositing scene be rendered.

The new command line to run the script now lokks like
"C:\Program Files\Blender Foundation\Blender\blender.exe" -b 1.blend -P video_editors_render_script.py [-- THIS_SCENE_NAME]

Yours sincerely,
Yann